# Optimization: eager syscalls 