TCP sockets should attempt the non-blocking read/write/accept in the main thread.

```
gpu ~/src/deno> ./profile_http.sh out/before/deno
Listening on 127.0.0.1:4500
Running 10s test @ http://localhost:4500/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   438.95us  432.04us   8.14ms   96.14%
    Req/Sec    13.12k     1.43k   15.13k    93.00%
  260978 requests in 10.01s, 12.69MB read
Requests/sec:  26080.64
Transfer/sec:      1.27MB

gpu ~/src/deno> ./profile_http.sh out/eager_read_only/deno
Listening on 127.0.0.1:4500
Running 10s test @ http://localhost:4500/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   396.49us  413.72us   8.73ms   96.20%
    Req/Sec    14.73k     1.31k   15.51k    96.00%
  293044 requests in 10.00s, 14.25MB read
Requests/sec:  29299.82
Transfer/sec:      1.43MB

gpu ~/src/deno> ./profile_http.sh out/eager_read_write/deno
Listening on 127.0.0.1:4500
Running 10s test @ http://localhost:4500/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   334.21us  447.99us  12.23ms   95.24%
    Req/Sec    19.28k     2.32k   21.70k    93.50%
  383787 requests in 10.01s, 18.67MB read
Requests/sec:  38352.23
Transfer/sec:      1.87MB

gpu ~/src/deno> ./profile_http.sh out/eager_read_write_accept/deno
Listening on 127.0.0.1:4500
Running 10s test @ http://localhost:4500/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   329.35us  452.19us  12.23ms   95.00%
    Req/Sec    19.94k     2.53k   22.16k    95.50%
  396920 requests in 10.00s, 19.31MB read
Requests/sec:  39686.59
Transfer/sec:      1.93MB
```