@ry 2018-10-20
I disagree with the idea that we should enforce ordering at this level of abstraction. Users should be able to issue an isolated read or write syscall if desired. Ping-ponging quickly /is/ what these (node, deno) runtimes do. Certainly we will provide high-level access, like readFile, but low-level access ought to be possible too.